+++ src/components/sidenav/index.js
@@ -1,5 +1,7 @@
 import AxSidenav from './Sidenav.vue';
 import AxSidenavLink from './SidenavLink.vue';
+import AxSidenavHeader from './SidenavHeader.vue';
+import AxSidenavFooter from './SidenavFooter.vue';
 
 import { use, registerComponent } from '../../utils/plugins';
 
@@ -7,6 +9,8 @@ const Plugin = {
   install(Vue) {
     registerComponent(Vue, AxSidenav);
     registerComponent(Vue, AxSidenavLink);
+    registerComponent(Vue, AxSidenavHeader);
+    registerComponent(Vue, AxSidenavFooter);
   },
 };
 
@@ -14,4 +18,4 @@ use(Plugin);
 
 export default Plugin;
 
-export { AxSidenav, AxSidenavLink };
+export { AxSidenav, AxSidenavLink, AxSidenavHeader, AxSidenavFooter };